Refactored `functionSetup` and `globalSettings` by splitting in dedicated library and a file in root
